Restructure olive-mcp-server into installable package
- Move mcp_server, tools, fetchers, and knowledge_base into
  olive_mcp_server package
- Add __init__.py and __main__.py for `python -m olive_mcp_server`
- Add console script `olive-mcp-server` in pyproject.toml
- Declare knowledge_base/*.json as package_data for pip installs
- Update imports in tests and scripts
- Verify `pip install -e .` and `python -m pytest tests -q` pass
