+++ olive-mcp-server/olive_mcp_server/__init__.py
@@ -1,0 +1,3 @@
+"""Olive MCP server package."""
+
+__version__ = "0.1.0"
