+++ olive-mcp-server/olive_mcp_server/__main__.py
@@ -1,0 +1,6 @@
+"""Entry point for `python -m olive_mcp_server`."""
+
+from olive_mcp_server.mcp_server import main
+
+if __name__ == "__main__":
+    main()
