+++ olive-mcp-server/olive_mcp_server/tools/hardware_guide.py
